v2: Ditch Query helpers. (#17)
This PR focuses around simplification of the tsplot package. Moving away
from something that was rather perscriptive in the type of Query that
you could make and towards a package that can accept a Time Series or
Time Series Itterator from the Google Metrics API and return a rendered
graph of those Time Series.